code businesscard

+++ scrm-service/src/main/java/com/huotu/scrm/service/repository/BusinessCardReposity.java
@@ -1,14 +1,15 @@
 package com.huotu.scrm.service.repository;
 
-import com.huotu.scrm.service.entity.BusinessCard;
+import com.huotu.scrm.service.entity.businesscard.BusinessCard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
  * 名片
  * Created by Administrator on 2017/7/11.
  */
-public interface BusinessCardReposity extends JpaRepository<BusinessCard , Integer>{
+public interface BusinessCardReposity extends JpaRepository<BusinessCard , Long>{
 
-    BusinessCard getByUserIdAndCustomerId( Integer userId , int customerId);
+    BusinessCard getByUserIdAndCustomerId( Long userId , Long customerId);
+
 
 }
